[CALCITE-7178] FETCH and OFFSET in EnumerableMergeUnionRule do not support BIGINT

+++ core/src/main/java/org/apache/calcite/adapter/enumerable/EnumerableMergeUnionRule.java
@@ -92,8 +92,8 @@ public class EnumerableMergeUnionRule extends RelRule<EnumerableMergeUnionRule.C
         inputFetch = sort.fetch;
       } else if (sort.fetch instanceof RexLiteral && sort.offset instanceof RexLiteral) {
         inputFetch =
-            call.builder().literal(RexLiteral.intValue(sort.fetch)
-                + RexLiteral.intValue(sort.offset));
+            call.builder().literal(RexLiteral.longValue(sort.fetch)
+                + RexLiteral.longValue(sort.offset));
       }
     }
